pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "repoverlay"
version = "0.1.0"
description = "Clone overlay repos and create symlinks for secrets/config management"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PyYAML>=6.0",
]

[project.scripts]
repoverlay = "repoverlay.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["src"]
